refactor: 불필요한 변수들 삭제(피드백 반영)

+++ services/comment.service.js
@@ -21,7 +21,7 @@ class CommentService {
   modifyComment = async (userId, content, commentId, password) => {
     const comment = await this.commentRepository.findComment(commentId);
     if (!comment) {
-      return { message: "없는 댓글입니다." };
+      return { code: 404, message: "없는 댓글입니다." };
     } else if (comment) {
       if (userId !== comment.userId) {
         return { code: 400, message: "댓글 작성자가 아닙니다." };
@@ -43,12 +43,14 @@ class CommentService {
 
   deleteComment = async (commentId, userId, password) => {
     const findComment = await this.commentRepository.findComment(commentId);
+    if (!findComment)
+      return { code: 404, message: "존재하지 않는 댓글입니다." };
     if (findComment.userId !== userId) {
       return { code: 400, message: "댓글 작성자가 아닙니다." };
     } else if (findComment.password !== password) {
       return { code: 400, message: "비밀번호를 확인해주세요." };
     } else {
-      const result = await this.commentRepository.deleteComment(commentId);
+      await this.commentRepository.deleteComment(commentId);
       return { code: 201, message: "댓글이 삭제되었습니다." };
     }
   };
